[Rust][CI] 修复 clippy -D warnings 阻塞 Rust CI
消除 legado-core/book/parser/db/net/js 在稳定版 clippy 下的告警（camelCase allow、map_or/is_multiple_of 等），使 cargo clippy -- -D warnings 可通过。

+++ rust/legado-db/src/repository/highlight_rule_repository.rs
@@ -69,7 +69,7 @@ impl<'a> HighlightRuleRepository<'a> {
                     ],
                 )
                 .map_err(|e| LegadoError::Database(format!("插入高亮规则失败: {e}")))?;
-            Ok(self.conn.last_insert_rowid() as i64)
+            Ok(self.conn.last_insert_rowid())
         }
     }
